Update DefaultPageLayout.php
This is a drop-in replacement for the ajaxSend() function. It will inject the csrf_token in every POST request done using ajax, and, if a status 403 is received from the server, it will retrieve the current csrf_token from the server and retry the request once.

This will require emhttp to trow a 403 error when csrf_token is wrong or missing.