Multi-language support
Required for CA to support multi-language.

Here's the story

When on multi-language, if the user selects "Dutch", then CA operates in Dutch.  However when installing a container from CA, the language on the Add Container screen will revert to English.

The workaround for that is for CA to specify the install URL as /Docker/Apps/AddContainer... instead of the current /Apps/AddContainer.

But, that change breaks the "Done" Button in returning to CA.  This change fixes that.  Doesn't appear to affect anywhere else in the UI where the Done button is also present, but @bergware might want to double check that