fix: session corruption via GUI boot mode
When accessing the webgui from Unraid's GUI boot mode,
the session does not persist and thus changes on every request.

This makes sessions ephemeral to each request, instead of persisting
e.g. user preferences or auth information.

The culprit of this seems to be `plugins/dynamix/include/Translations.php`,
specifically the second block of session handling logic: "remove empty
and temporary session files".

In some edge case triggered only in GUI boot mode (possibly a race condition),
this will delete the current, valid session file.

In the interest of backwards compatibility and preserving existing auth
behavior, I opted not to delete this code (which does resolve the
issue/symptom) and instead add a self-healing hack to the top of
`DefaultPageLayout`.

Resolving this session corruption allows the webgui to access the
Unraid API in GUI mode using its session cookie.